mcp_server/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
sqlalchemy==2.0.20
alembic==1.12.0
psycopg2-binary==2.9.7
asyncpg==0.28.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
python-dotenv==1.0.0
pgvector==0.2.1
openai==0.28.0
langchain==0.0.267
cryptography==41.0.3
pydantic-settings==2.0.3
authlib==1.2.1
